move `define_callback_for_cleaning_cache` to  cache_service
