+++ lib/active_model_cachers/active_record/extension.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
-require 'active_model_cachers/column_value_cache'
 require 'active_model_cachers/active_record/attr_model'
 require 'active_model_cachers/active_record/cacher'
 require 'active_model_cachers/hook/dependencies'
@@ -24,7 +23,7 @@ module ActiveModelCachers
         with_id = true if expire_by.is_a?(Symbol) or query.parameters.size == 1
         expire_class, expire_column, foreign_key = get_expire_infos(attr, expire_by, foreign_key)
         if expire_class
-          define_callback_for_cleaning_cache(expire_class, expire_column, foreign_key, on: on) do |id|
+          service_klass.define_callback_for_cleaning_cache(expire_class, expire_column, foreign_key, on: on) do |id|
             service_klass.clean_at(with_id ? id : nil)
           end
         end
@@ -70,35 +69,6 @@ module ActiveModelCachers
         Cacher.define_cacher_method(attr, attr.primary_key, service_klasses)
         ActiveSupport::Dependencies.onload(attr.class_name) do
           service_klasses << cache_self
-        end
-      end
-
-      @@column_value_cache = ActiveModelCachers::ColumnValueCache.new
-      def define_callback_for_cleaning_cache(class_name, column, foreign_key, on: nil, &clean)
-        ActiveSupport::Dependencies.onload(class_name) do
-          clean_ids = []
-
-          prepend_before_delete do |id, model|
-            clean_ids << @@column_value_cache.add(self, class_name, id, foreign_key, model)
-          end
-
-          before_delete do |_, model|
-            clean_ids.each{|s| clean.call(s.call) }
-            clean_ids = []
-          end
-
-          after_delete do
-            @@column_value_cache.clean_cache()
-          end
-
-          on_nullify(column){|ids| ids.each{|s| clean.call(s) }}
-
-          after_commit ->{
-            changed = column ? previous_changes.key?(column) : previous_changes.present?
-            clean.call(send(foreign_key)) if changed || destroyed?
-          }, on: on
-
-          after_touch ->{ clean.call(send(foreign_key)) }
         end
       end
     end
